Moved auth to feature module with lazy loading
