+++ src/app/auth/auth.component.html
@@ -1,0 +1,12 @@
+<div class="container-fluid">
+    <div class="main-layout">
+      <header class="header">
+        <a href="#">
+          <img class="app-logo" src="https://accounts.app.fyle.in/app/router/assets/images/logo/fyle_logo_light.png"
+            alt="Fyle">
+        </a>
+        <h3 class="app-name">Fyle QBO Integration</h3>
+      </header>
+      <router-outlet></router-outlet>
+    </div>
+  </div>
