+++ src/app/auth/auth.component.spec.ts
@@ -8,8 +8,9 @@ describe('AuthComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [AuthComponent],
-    }).compileComponents();
+      declarations: [ AuthComponent ]
+    })
+    .compileComponents();
   }));
 
   beforeEach(() => {
