+++ src/app/auth/auth.module.ts
@@ -1,0 +1,23 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { AuthRoutingModule } from './auth-routing.module';
+import { AuthComponent } from './auth.component';
+import { LoginComponent } from './login/login.component';
+import { CallbackComponent } from './callback/callback.component';
+import { LogoutComponent } from './logout/logout.component';
+
+
+@NgModule({
+  declarations: [
+    AuthComponent,
+    LoginComponent,
+    CallbackComponent,
+    LogoutComponent
+  ],
+  imports: [
+    CommonModule,
+    AuthRoutingModule
+  ]
+})
+export class AuthModule { }
