+++ src/app/core/guard/auth.guard.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, Router } from '@angular/router';
-import { AuthService } from '../../components/auth/auth.service';
+import { AuthService } from '../services/auth.service';
  
 @Injectable({
   providedIn: 'root'
